+++ dotnet/derftx.service/program.cs
@@ -1,0 +1,8 @@
+using derftx.service;
+
+var builder = WebApplication.CreateBuilder(args);
+var app = builder.Build();
+
+app.Run(Server.Root);
+
+app.Run();