+++ dotnet/derftx.service/server.cs
@@ -1,0 +1,10 @@
+namespace derftx.service
+{
+  public static class Server
+  {
+    public static Task Root(HttpContext context)
+    {
+      return context.Response.WriteAsJsonAsync(new { runtime = ".NET Core"});
+    }
+  }
+}